[Task 53] Implement Nestests for CPU (#55)
* updated NESCPU to use bus memory

* updated unit tests to use bus memory

* Implemented NESTEST unit tests to test all the instructions.

* updated methods to pass unit tests

* Updated and added unit tests for addressing modes. Fixed addressing modes.

* added comment - link to nesdev.com addressing mode explanation.

* Comments added

+++ WojoNES/Core/Cartridge/Cartridge.swift
@@ -7,7 +7,7 @@
 import Foundation
 
 protocol Cartridge: AnyObject {
-    func write(data: Data, address: UInt16)
-    func read(address: UInt16) -> UInt8
+    func write(data: Data, address: Int)
+    func read(address: Int) -> UInt8
     func reset()
 }
